requirements.txt
discord==2.3.2
python-dotenv==1.0.1
pillow==10.2.0
requests==2.31.0
replicate==0.25.1